+++ src/main/java/ru/deloom/jpaservice/repositories/KeywordsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ru.deloom.jpaservice.entities.Keywords;
 
-public interface KeywordsRepository extends JpaRepository<Keywords, Long>{
+public interface KeywordsRepository extends JpaRepository<Keywords, Integer>{
 	List<Keywords> findByName(String name);
 }
